Add use cases
